refactor(federation): rename schema-loom.ts
